move rules and analysis out
