+++ src/rules.py
@@ -1,5 +1,6 @@
 """Rules."""
 import streamlit as st
+from preprocess.pipeline import run
 
 
 def toggle_rules():
